+++ mcontext/multiple_context.go
@@ -32,19 +32,19 @@ func NewMultipleContext(c ...context.Context) *MultipleContext {
 
 // Listen is required to called before using the MultipleContext object.
 // Dispose is required to called when the object is not need.
-func (this *MultipleContext) Listen() {
-	this.runner.Mark()
-	go this.running()
+func (mc *MultipleContext) Listen() {
+	mc.runner.Mark()
+	go mc.running()
 }
 
 // Dispose is called with Listen in pair.
-func (this *MultipleContext) Dispose() {
-	this.runner.CloseWait()
+func (mc *MultipleContext) Dispose() {
+	mc.runner.CloseWait()
 }
 
 // Hit return the context had Done.If there are no context Done, nil will be returned.
-func (this *MultipleContext) Hit() context.Context {
-	v := this.hit.Load()
+func (mc *MultipleContext) Hit() context.Context {
+	v := mc.hit.Load()
 	if v == nil {
 		return nil
 	}
@@ -52,45 +52,45 @@ func (this *MultipleContext) Hit() context.Context {
 }
 
 // Done is the implement of context.Context interface
-func (this *MultipleContext) Done() <-chan struct{} {
-	return this.c.Done()
+func (mc *MultipleContext) Done() <-chan struct{} {
+	return mc.c.Done()
 }
 
 // Value is the implement of context.Context interface
-func (this *MultipleContext) Value(key interface{}) interface{} {
-	return this.c.Value(key)
+func (mc *MultipleContext) Value(key interface{}) interface{} {
+	return mc.c.Value(key)
 }
 
 // Deadline is the implement of context.Context interface
-func (this *MultipleContext) Deadline() (time.Time, bool) {
-	c := this.Hit()
+func (mc *MultipleContext) Deadline() (time.Time, bool) {
+	c := mc.Hit()
 	if c != nil {
 		return c.Deadline()
 	}
-	return this.c.Deadline()
+	return mc.c.Deadline()
 }
 
 // Err is the implement of context.Context interface
-func (this *MultipleContext) Err() error {
-	c := this.Hit()
+func (mc *MultipleContext) Err() error {
+	c := mc.Hit()
 	if c != nil {
 		return c.Err()
 	}
-	return this.c.Err()
+	return mc.c.Err()
 }
 
-func (this *MultipleContext) running() {
+func (mc *MultipleContext) running() {
 	defer func() {
-		this.cancel()
-		this.runner.Done()
+		mc.cancel()
+		mc.runner.Done()
 	}()
 	cases := []reflect.SelectCase{
 		{
 			Dir:  reflect.SelectRecv,
-			Chan: reflect.ValueOf(this.runner.Quit()),
+			Chan: reflect.ValueOf(mc.runner.Quit()),
 		},
 	}
-	for _, c := range this.cc {
+	for _, c := range mc.cc {
 		cases = append(cases, reflect.SelectCase{
 			Dir:  reflect.SelectRecv,
 			Chan: reflect.ValueOf(c.Done()),
@@ -100,5 +100,5 @@ func (this *MultipleContext) running() {
 	if chosen == 0 {
 		return
 	}
-	this.hit.Store(this.cc[chosen-1])
+	mc.hit.Store(mc.cc[chosen-1])
 }